Part of functionality was moved from User app to Github Integration app
