+++ sandbox/user/templates/user/profile.html
@@ -10,10 +10,10 @@
             <a href="{% url 'core:projects' %}">My Projects</a>
         </li>
         <li class="list-group-item">
-            <a href="{% url 'user:repositories' %}">My Repositories</a>
+            <a href="{% url 'github_integration:repositories' %}">My Repositories</a>
         </li>
         <li class="list-group-item">
-            <a href="{% url 'user:github_repositories' %}">My GitHub Repositories</a>
+            <a href="{% url 'github_integration:github_repositories' %}">My GitHub Repositories</a>
         </li>
     </ul>
 {% endblock %}
